Port to pyexiv2 0.3.
GottenGeography now depends on pyexiv2 0.3. This port was a little bit confusing for me, because I was expecting pyexiv2 to support fractions.Fraction objects natively, but in practise it seems as though it gives me Fraction objects when loading from disk, but still expects me to give it pyexiv2.Rational objects when I'm writing back to disk.

Anyway, this commit brings the test suite fully into line, 100% pass, with just a little bit of Fraction/Rational juggling.

Ready for release? Maybe.