Change the way dependencies are checked.
They were in __init__.py before because I liked the way that that
implicitly tested everything on launch. But now that it's been
changed to only do the dependency testing when the program is
being run, might as well just move the dep tests into the app
directly.

This also adds a check for libchamplain 0.12.0