Issue #41: Overhaul GSettings with overrides.
This commit introduces bind_with_convert(), which is a pure-python
reimplementation of g_settings_bind_with_mapping that's missing
from the python API for some reason.

The first use of this is to store the GdkColor of the GPS track,
so that's no longer in GConf.

With many thanks to John Stowers of gnome-tweak-tool for most of
this code.