Use dark GTK theme, and larger dialog icons.
I feel the dark theme is justified because it's a photo editing app,
and the primary widget is the map, so the darker theme detracts less
from the map itself.

This commit also increases the size of dialog icons, which I did
primarily to make the app icon more visible on the About dialog.
This does also impact the close dialog and I'm ok with that.