Overlay clock offset spinbuttons on top of map.
This abolishes GtkStatusBar.

TODO: Think of an unobtrusive way to display status. I'd like the
"unable to open" error message to be displayed over the map similar
to the way gedit shows error messages directly in the text editing
area (ie, not as a popup error dialog).

Also TODO: We seem to have lost the etched shadow around the map.
d'oh!