Tweak icon sizes.
This commit causes the About dialog to load the application icon
at a specific (large) size, meaning I no longer have to override
the default 'gtk-dialog' size to be larger, which means that the
warning icon on the Quit dialog is now back to being normal size.

This also means that I can use the standard gtk-dialog size for
the GtkInfoBar, again increasing consistency overall with other
GNOME applications.