Use less magic and more strong typing for history, window-size.
This splits up GSettings.set magic into distinct GSettings.set_history
and GSettings.set_window_size, which means no more guessing
GVariantTypes based on Python data types. This is in anticipation
of future expansion of the use of GSettings and GVariants.

This commit also changes the GVariantType of the history stack from
an array of arrays of floats ('aad') to an array of float,float,int
tuples ('a(ddi)'). This makes the history stack a little more rigid,
a little more predictable, and a little more correct since the zoom
level was never meant to be a float.