Add some support for camera serial numbers.
On a quick test with some files I had lying around, this worked with
a Nikon D300s, but not a D200, and neither did it work for a Canon A70
or A590. But it's still better than nothing (The absolutely *enormous*
group of people who own more than one Nikon D300s, don't synchronise
the clocks, and use my program will be thankful).