Restore the timezone lookup ability.
The way it's currently implemented is that the camera stores the
timezone directly into GSettings every time a photo is placed onto
the map (whether it be automatically or manually), and then if the
timezone method is changed to lookup, it consults the GSettings key
to determine where the camera was last used.

There are some subtleties here... if you open photos first, then GPX,
the GPX simply looks at the loaded cameras and tells them 'hey, this
is the timezone here' and then the photos are placed correctly. If
you open the GPX before photos, the photos don't have any way of
communicating with the GPX to ask what the timezone is, so they
default to use the system timezone. I put in a check for this at
the end of GottenGeography.open_files so that the timezones will
be recalculated correctly, but it's not the sexiest code ever.

Thankfully the discovered timezone is persisted in GSettings so this
is only an issue the very first time that a camera is loaded (and again
every time you start using the camera in a new timezone and change
the camera's clock to match).

This works I guess, but the results are undefined in the event that
your GPX track crosses a timezone boundary. Because of this, I *highly*
recommend *always* keeping your camera clock synchronised to your
computer's clock and then just use 'system' instead of 'lookup'.