Initial schema for #61.
I think that we will need to have both a global and a per-trackfile
color setting, but we can still ditch the preferences dialog and
just have a little color chooser at the top of the trackfile list.
This way users can still choose a whatever color is pleasing to them
for their newly loaded trackfiles, and then they can tinker with
each track independantly if they like.