Issue #61. Hookup timezone lookup, and better UI for default colors.
I removed the GtkColorButton that let the user explicitely choose
the global default. The new way is much simpler, where the global
default is simply determined by whichever color was most recently
chosen for whatever file. In effect, trackfiles opened for the
very first time will take on the color of the most-recently-changed
trackfile, and then if the user changes this color, that new color
becomes the default for all future trackfiles, etc.