Make the Apply button magical ;-) Closes #70.
Ok, I will try to explain the magic here. Basically I have streamlined
Select All and Apply into one button, but smarter. Now, when you click
the apply button, it scans all the loaded photos for ones that do
not have a manual position set, and sets those ones to the map center.

The idea is that you can use drag & drop to manually position some
photos, and then press the apply button to lazily slap all the remaining
photos onto the map center.

And the best part is that after you click it, it disappears, because
there's then nothing left for it to do.

The primary use-case is that the user would just load a ton of photos,
and then they can click the apply button to batch-apply ALL of the
photos onto map center, making for a very fast and streamlined "batchy"
operation. Drag & drop can then be used either before or after the apply
button has been pressed to fine-tune individual photos.