UI *overhaul*, Closes #86.
Ok, here's the big one...

* Search bar has been pushed up into the main toolbar, creating more
room for a larger map widget. This is because the map is the main
widget that is always visible, so the primary toolbar is mostly
widgets that act on the map.

* New secondary toolbar at the bottom of the Photos tab that only
contains buttons that act on photos, and is hidden when the photos
are hidden. Contains Close, Revert, Save All, and Apply. Thanks
to these buttons being placed once again on a GtkToolBar, their
tooltips are working again so that should help users with clarity.

* While I was at it, I totally reworked the GtkBuilder structure,
ditching a series of nested GtkVBoxes and GtkHBoxes in favor of more
simply putting everything onto one big GtkGrid and then positioning
them accordingly. Removes a layer of indentation from almost the
whole file...