Convert to using Nose for tests. Issue #33.
This offers an *enormous* benefit to readability and maintainability,
but there is a minor drawback in that the standard Python 'assert's
give a less helpful error message when they fail. I'm ok with this,
though, all I have to do is add a print statement before a failing
assertion and then I see what's going on.

Note that the prodigious amount of imports from the old testsuite.py
translates into just a handful of imports per file. Makes things way
more readable overall.