Merge metadata class into TrackFile class.
metadata only stored the timestamps of the first and last TrackFile,
so it makes sense that TrackFile can keep track of that itself.

This commit also moves auto_timestamp_comparison from common.py
to photos.py, partly because that's the only place it's used,
and partly because it was necessary to avoid a cycling import
infinite loop.