Make accessing GtkBuilder much more pythonic and sexy.
This throws out all that random get_obj('junk') junk that was hard
to read and not very meaningful. In it's place, we now have Widgets,
which is an instance of GtkBuilder, and has the ability to access
the defined widgets either as instance attributes (Widgets.gtk_builder_id)
or alternately as Widgets['gtk_builder_id'] for those places where
the widget id gets calculated at runtime.

I consider this to be an enormous improvement in readability primarily
because it gives us access to all the widgets with just a single period,
as opposed to two brackets and two quotes necessary before.

There are a couple limitations though. First off, if you use
hyphens in the GtkBuilder object id's, then you'll have to use
Widget['hyphenated-id'] because Widget.hyphenated-id is invalid
(Python will think you're trying to subtract id from Widget.hyphenated).

Also, we do run a small risk of namespace collisions between the actual
instance attributes of the GtkBuilder object and the object ids defined
in the GtkBuilder xml, but if you use meaningful ids you should be fine.
Consult dir(Widgets) for a list of what not to use as object ids.