Restore the ability to manually position photos.
The photo.manual attribute indicates that the user has manually
positioned the photo, overriding any auto timestamp logic from the
loaded GPX data.

This allows, eg, the user to drag a ChamplainLabel off of the GPX
track, and then change the camera offset, and have the other photos
slide up and down the GPX track without the manually positioned
photo snapping back on to the GPX track. We should be making
an effort to honor and preserve anything the user does by hand,
as it's quite frustrating to place a label delicately and then
have the program lose that data. It's equivalent to the power
going off before you've saved your word processing document. ;-)