Add another secret GSetting for changing the thumbnail size.
I should probably find a way to expose some UI for this one. People
with smaller screens may want smaller thumbnails as I discovered
when I lowered my screen resolution to record that screencast.

Even better would be a way to specify the thumbnail size in a
resolution-independant kind of way.