Don't update the window titlebar more than once per 60 seconds.
Map navigation had become quite jerky because we were doing geoname
lookups every time the map moved even slightly. The cache had
helped a small amount with this but we were suffering from excessive
cache misses.

I might consider making the timeout less than 60 seconds but for now
I've set it quite high because it's nice to be able to move the map
smoothly, and the titlebar showing accurate geoname information is
not a critical feature.