Simplify window title logic with bindings.
Drops the handler func in favor of a direct binding. This has the
side effect that the titlebar no longer bears the name of the app,
but that doesn't bother me because the Gnome Shell top bar still
says it, and so does the about dialog.

Also reduces the delay from 60 seconds to 10. 10 seconds is plenty
to keep the map feeling responsive while not letting the titlebar
get too stale.