Implement TrackFile.query_all_timezones()
It scans all the loaded GPX/KML and if it finds that they are all
in the same timezone, it reports that zone.

Before this, the app would just grab a random trackfile and pick
it's zone, which wasn't very reliable or useful. This new method
admits that if the user has trackfiles spanning multiple zones,
then we simply can't know what timezone their camera is set to. But,
if they all match, then it's likely that that's the timezone they're
in.