Fix "Jump to" button.
The behavior of valid_coords() has changed... it used to mean if
the photo was placed on the map because the coords would default to
None, which is invalid. But now thanks to the gobject properties
they default to 0, which are valid coords... so we need to use
photo.positioned to indicate whether the photo has coords or not.