Merge Coordinates.valid_coords() into Coordinates.positioned property.
It didn't make sense to maintain a separate _positioned attribute
when it could be calculated fairly easily whether the photo had
been positioned or not.

After some testing I determined that it's effectively impossible to
place a photo at 0,0 on the map. You have to zoom right in to max, and
fiddle with it quite a bit. I was able to get longitude=0 but I
actually could not achieve latitude=0. Therefore if a photo is actually
at precisely 0,0 that is a pretty good indication that it actually
hasn't been tagged yet and is just sitting at the default location.