Fix a bug that was causing 'notify::geoname' to be emitted twice.
This was causing redundant calls to loaded_photos.set_value, ie,
updating the text in the liststore twice.

Also, we were emitting 'notify::geoname' before we'd actually set
the _geoname property, so on the first call it wasn't even possible
to access the updated geoname information. Now we only emit the
signal once, only when it actually changes, and the information
is updated correctly at that time.