Don't update the ListStore unless the geoname has actually changed.
'is not' doesn't work on strings, it will be True for different
but equal strings. Now we are truly only emitting 'notify::geoname'
when the geoname actually changes, not every single time that it's
replaced with itself (effectively, the previous behavior would
emit notify::geoname every time there was a @memoize cache miss,
even if the newly looked up geoname was identical to the old).