Simplify Coordinates.positioned property.
It no longer validates that the lat/lon coords are within range,
because it turns out that GObject properties actually validate
the ranges themselves, ie, it's not possible to actually set an
invalid value in the first place. However it still needs to return
False when the coords are precisely 0.0,0.0 (indicating the default
state), so I've simplified that logic significantly.