Move GottenGeography.status_message to Widgets singleton.
It's a more natural fit here, because it was the last thing that
GottenGeography.load_gpx_from_file referenced `self` for, and
it also makes it easier to display warnings and errors from anywhere
in the source, instead of just in app.py.

Also, app.py has now shockingly dropped below 300 lines of code.