Fix bug that made reverting photos difficult.
For some reason you had to click revert twice for it to fully take
effect. This fixes that, reverting works correctly. There was also
a related bug where reverted photo labels would move to 0,0 for a
second before vanishing, this fixes that as well.

This commit also changes the behavior of the apply button, it
removes all of the magic (which wasn't working very well at all)
and brings it back to a more predictable behavior of simply taking
only the selected photos, and placing them at map center. The main
improvement here is that it doesn't disable itself after use, so
you can apply the same photos over and over again until you get
the location just right.