Fix an unhandled exception when unloading duplicate data.
If you have two overlapping GPS traces, the `points` dict will
only contain the points of the most-recently-loaded trace. If
you then try to unload both of those traces, the second one
will result in KeyErrors as it tries to unload data that
was already unloaded by the previous trace unloading.

The solution is instead to clear the `points` dict and then
rebuild it from scratch using whichever traces are leftover.
This way we don't have to guess which trace provided which data,
and now the `points` dict is guaranteed to be in sync with the
loaded traces.