Simplify/unify some altitude logic among file formats.
I pushed the call to float() into Polygon.append_point, partly
because now not every single TrackFile subclass has to call it
itself, but also because this new try statement makes it more
robust in the event that the altitude is missing or corrupt.
i.e., there was an unhandled exception if the altitude field
contained letters or was blank, but not any longer.