*Major* performance overhaul.
I was recently tagging a batch of over 100 photos and found myself
quite disappointed with the performance, particularly with the
camera clock offset slider being quite sluggish to move the photos
around the trace. So, I fired up cProfile and identified a number
of bottlenecks, which have been significantly reduced by this
commit.