Faster string comprehensions.
Looks like I made a mistake the first time I profiled these. Here
are the correct performance results, using timeit:

", ".join((s for s in ["a", "b", None, "c"] if s))  1.77 usec
", ".join(s for s in ["a", "b", None, "c"] if s)    1.77 usec
", ".join(s for s in ("a", "b", None, "c") if s)    1.67 usec
", ".join([s for s in ["a", "b", None, "c"] if s])  0.73 usec
", ".join([s for s in ("a", "b", None, "c") if s])  0.60 usec

So the fastest combination is to start with a tuple, then use
list comprehension on it. The other way around takes 3x as
long to execute.