16% speedup in the KML parser.
The biggest change is that I've moved it to a two-pass approach in
which the XML is parsed into memory first, and then the points
are placed onto the map second. The reason this is faster is that
originally, the logic for checking how many complete when/coord
pairs were present happened in KMLFile.element_end, which was
quite slow and happened once for every single closing xml tag
(ie, twice for every when/coord pair).

The new approach cuts that out and puts it into a new second pass,
where it only happens once at the very end, and is quite a bit
faster at it, too. I also make use of deques, for faster access
to popleft().