Major simplification of the memoizer.
First, I started using functools.wraps() in order for the memoizer
function to better masquerade as the class/method/function being
memoize'd, without that ugly __dict__ hack I had before.

However, this broke staticmethods all over again because they just
don't like being on a class that gets wrapped (which was the reason
for most of that __dict__ ugliness in the first place). This time,
though, I solve that by overriding the builtin staticmethod with
my own implementation that works a lot better than the original,
in a much less hacky way.

It's unfortunately much more verbose than the old @memoize
implementation, but I feel it's a much cleaner solution.

This commit also alleviates the necessity for classes to define
their own `instances` dict, except in the case of TrackFile where
all the TrackFile subclasses need to share. I've also renamed
`instances` to `cache`, and defined a new `instances` as
`cache.viewvalues()`, which I find makes the code a lot more
intuitive to read.