More doctests, especially for @memoize'd classes.
One of the benefits of using functools.wraps in the previous commit
is that we can now have doctests in a @memoize'd class. Unfortunately
it still doesn't work in any of the methods, but at least it does
on the classes themselves.

But in addition to that this commit also introduces a lot more
doctests in general, @memoized or not. Some of the nose tests became
redundant, so I deleted some of those as well.