Convert Coordinates.pretty_coords() to a GObject property.
This gives us a couple advantages: It means that we can use a
Binding in order to display the map center coordinates, re-using
the same 'center' Coordinates instance that we were already using
to set the Window title.

This also means that actor.py no longer imports from gpsmath.py,
which means that format_coords only gets called from one place,
which means we can inline it and halve the number of function
invocations it takes to draw map center coordinates onto the map.

Now we're thinking with GObjects!